refactor(#730): extract the submitter machine actors to own files
